+++ packages/cw1/src/msg.rs
@@ -1,4 +1,3 @@
-use schemars::JsonSchema;
 use std::fmt;
 
 use cosmwasm_schema::cw_serde;
@@ -7,7 +6,7 @@ use cosmwasm_std::{CosmosMsg, Empty};
 #[cw_serde]
 pub enum Cw1ExecuteMsg<T = Empty>
 where
-    T: Clone + fmt::Debug + PartialEq + JsonSchema,
+    T: Clone + fmt::Debug + PartialEq,
 {
     /// Execute requests the contract to re-dispatch all these messages with the
     /// contract's address as sender. Every implementation has it's own logic to