+++ packages/cw20/src/coin.rs
@@ -1,16 +1,16 @@
 use cosmwasm_schema::cw_serde;
-use cosmwasm_std::{Addr, Uint128};
+use cosmwasm_std::{Addr, Uint256};
 use std::fmt;
 
 #[cw_serde]
 pub struct Cw20Coin {
     pub address: String,
-    pub amount: Uint128,
+    pub amount: Uint256,
 }
 
 impl Cw20Coin {
     pub fn is_empty(&self) -> bool {
-        self.amount == Uint128::zero()
+        self.amount == Uint256::zero()
     }
 }
 
@@ -23,12 +23,12 @@ impl fmt::Display for Cw20Coin {
 #[cw_serde]
 pub struct Cw20CoinVerified {
     pub address: Addr,
-    pub amount: Uint128,
+    pub amount: Uint256,
 }
 
 impl Cw20CoinVerified {
     pub fn is_empty(&self) -> bool {
-        self.amount == Uint128::zero()
+        self.amount == Uint256::zero()
     }
 }